Add files via upload
I tried to post these yesterday but i don't think it worked.  I messed around at work and coded something similar to our assignment.  I was able to use a table at work and show the contents of the table in a grid.  I added edit, delete and add buttons and was able to get the id to pass down to javascipt functions.  I did this just in case you rand into any problems developing the front end.